Fix Error Causing Voting to Fail
Voting using `.vote()` checks if the post is archived, and if it is, it raises the `VotingInvalidOnArchivedPost` error.

However, prior to this fix, it checked if the `net_rshares` of the post was 0, which would also fire if the post did not have any votes on it yet. This fix allows posts that aren't archived but have 0 votes to be voted on.