Rewrite dictionary.py so it doesn't make grep explode
When grepping around the codebase, often a result will be
found in dictionary.py and then it attempts to display a
300kb long line, which makes it hard to tell whether there
are relevant results before/after.

This patch refactors dictionary.py to put each word on its
own line, and then replaces the resulting newlines with
commas to retain compatibility.